# Conflicts:
#	src/main/java/dev/foxikle/webnetbedwars/commands/DebugCommand.java

+++ src/main/java/dev/foxikle/webnetbedwars/listeners/MoveListener.java
@@ -17,6 +17,7 @@ public class MoveListener implements Listener {
 
     @EventHandler
     public void onMove(PlayerMoveEvent event){
+        if(!plugin.getGameManager().STARTED) return;
         if(event.getTo().y() <= -40){
                 plugin.getGameManager().kill(event.getPlayer(), null, EntityDamageEvent.DamageCause.VOID);
             event.setCancelled(true);
